Split coupling.h into coupling.cc and coupling.inst
