+++ source/non_matching/CMakeLists.txt
@@ -16,10 +16,12 @@
 INCLUDE_DIRECTORIES(BEFORE ${CMAKE_CURRENT_BINARY_DIR})
 
 SET(_src
+  coupling.cc
   immersed_surface_quadrature.cc
   )
 
 SET(_inst
+  coupling.inst.in
   )
 
 FILE(GLOB _header
